requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-pptx==0.6.23
python-multipart==0.0.9
Pillow==10.4.0
pytest==8.3.3
httpx==0.27.2
